factupid/requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.13.4
black==25.9.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
Django==5.1.4
django-admin-material-dashboard==1.0.21
django-autocomplete-light==3.12.1
django-bootstrap-modal-forms==3.0.4
django-material-admin==1.8.6
django-nested-admin==4.1.1
django-session-timeout==0.1.0
django-soap==0.1.9
django-tabbed-admin==1.0.4
django-view-breadcrumbs==2.5.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
fonttools==4.59.0
greenlet==3.2.4
html5lib==1.1
idna==3.10
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mypy_extensions==1.1.0
numpy==2.2.3
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pdfkit==1.0.0
pillow==11.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.1
pyphen==0.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-monkey-business==1.1.0
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
reportlab==4.2.5
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.23.1
satcfdi @ git+https://github.com/SAT-CFDI/python-satcfdi@a71cea5ffc26f94e02e74ed93e464613b9272541
six==1.17.0
soupsieve==2.7
spyne @ git+https://github.com/arskom/spyne.git@f105ec2f41495485fef1211fe73394231b3f76e5
SQLAlchemy==2.0.43
sqlparse==0.5.3
suds==1.1.2
suds-py3==1.4.5.0
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
xlrd==2.0.2
xlsxwriter==3.2.5
xmltodict==0.9.2
zopfli==0.2.3.post1
